用户短信的验证功能

+++ src/main/java/com/miaosha/service/model/UserModel.java
@@ -48,7 +48,7 @@ public class UserModel {
         return gender;
     }
 
-    public void setGender(Byte gender) {
+    public void setGender(Integer gender) {
         this.gender = gender;
     }
